requirements.txt
gradio
google-generativeai>=0.8.3
python-dotenv>=1.0.0
Pillow>=10.0.0
Pillow>=10.0.0
openai>=1.0.0
beautifulsoup4
requests
python-whois